feat: implement kubernetes node audit controller
Clean up invalid nodes (from the Kubernetes perspective) from the cluster if they are no longer part of a cluster from the perspective of Omni.

Introduce a new intermediate resource, `ClusterKubernetesNodes`, containing the ordered list of nodes in a cluster.

Create the resource, `KubernetesNodeAuditResult`, to provide a summary on the latest deleted nodes from the Kubernetes cluster.

Closes siderolabs/omni#9.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>