fix: properly select schematics for machine set and machine levels
The controller code has the bug in the loop: the last fallback was
picking resources which have cluster label, but the thing is that _all_
of them have it.
Make it select only the resources which have _only_ cluster label.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>
(cherry picked from commit 8aa6a6af152fb02475d2704c63511eec771dd35d)