test: fix the flakiness in the resourcelogger test
The resourcelogger unit test was flaky - in very rare occasions when there is a change in second between the creation and update time of the resources, it logged a different diff than the expected one.

Solve it by waiting a bit, so the `updated` field will always appear in the diff, as there's more than a second difference between the creation and update now.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>