fix: do not allow adding ISO, PXE nodes running different Talos version
Adding them to cluster will no longer work. Validation and the UI won't
allow that.

We can not upgrade the node which doesn't have Talos installed yet.
So the config generated for a different version of Talos will make this
node never join the cluster. That's why we disable that.

Fixes: https://github.com/siderolabs/omni/issues/43

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>