fix: start watch before delete in `omnictl delete`
Start the watch before sending the destroy call to not miss events.

Add support for passing more than a single resource to the call, e.g., `omnictl delete link link-1 link-2 link-3`.

Additionally, bring back the "WatchKind" behavior when `--all` or `--selector` is used, or when the number of resources to be deleted is above a certain size.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>