feat: improve auth flow when user email is not verified
If the user has their email not verified, instead of failing with a generic error message of "invalid JWT", print an error message asking user to verify their email and try again.

In Auth0 mode, if the JWT validation has failed on the backend at the moment of clicking "Login", get a new ID token from Auth0 on the next click. This way, the user will not have to reload the page after validating their email - they can simply click "Login" again to get in.

Part of siderolabs/omni#114.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>