feat: implement labels completion for clusters and machines
New virtual resource `LabelsCompletion` can now be used to query all
labels.
And it works for any resources in the default namespace.
To get the completions for any resource it is enough to pass the full
resource type as the completions ID.

Labels input is reusable and can be enable for any views we need.
Both mouse and keyboard control works for the completion selection.
Labels now appear as the same boxes as they are displayed in the
machines and cluster lists.

Label completion search looks both in keys and values until you add `:`
to the line. Then it looks for keys that contain whatever is before `:`
with values that contain whatever is after `:`.

Fixes: https://github.com/siderolabs/omni/issues/24

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>