fix: do not block machine config updates if loadbalancer is down
This check doesn't make sense for the config updates.
We are trying to address the case when the config has broken the API
server and you try to revert it. All api servers are down and you can
not revert it because all api servers are down.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>
(cherry picked from commit a0d02ea20b2ebcc6556e80afe04cb22392afe561)