feat: show service events on the node overview page
Also pull the service list from the `MachineService.ServiceList` API as
it has more data than the resource and has services which haven't
started yet.

The list is dynamically updated by subscribing on the machine events and
filtering the events of type `ServiceStateEvent`.

Additionally make the sidebar service list dynamically update for
consistency.

Fixes: https://github.com/siderolabs/omni/issues/21

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>