fix: ignore MachineStatus events timestamps as they're not reliable
The timestamp can jump back and forths if the Talos node time is out of
sync, so we might ignore proper events.

So instead of reading the timestamp from the event, use Omni server time
at the point when the event comes to the system.

In the worst case we'll see the node going through several phases
quickly.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>