fix: fix config generation failure due to secure boot status
A change in the `SecureBootStatus` in the `MachineStatus` did not cause a new config to be generated. This was caused by the `ClusterMachineConfigController` not taking the `MachineStatus` into the calculation of the hash from the inputs (`UpdateInputsVersions`) to determine if it should generate a new config or not.

Modify the existing `MachineConfigGenOptions` resource to hold the required information to build the install image. Instead of using `MachineStatus+ClusterMachineTalosVersion` to build the install image URL, use only `MachineConfigGenOptions` resource.

Additionally, add a migration to pre-populate the new fields in the `MachineConfigGenOptions` resources, so that they will not be modified on the next Omni upgrade. Also update the input versions of the `ClusterMachineConfig` to not contain `ClusterMachineTalosVersion` anymore, so that we won't trigger a config re-apply on the next Omni upgrade.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>