fix: pass through the `talosctl -n` args if they cannot be resolved
We were not correctly checking if the nodes passed via `talosctl --nodes` were resolved before replacing the target in the GRPC metadata with the resolved ones.

We were handling a single node in the metadata correctly, but not doing the same handing for multiple nodes. This PR fixes that.

This PR simplifies that logic.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>