fix: redo EtcdManualBackupShouldBeCreated
The main problem with this test is that two backups can be created within the same second. It doesn't happen often,
but when it happens it results in the situation where we have "newer" backup overwriting the "old" one. We can protect
against that for local storage, but S3 doesn't support "If-None-Match" header yet.

In practice, it shouldn't matter if manual backup overwrites automatic one because we are interested in what's inside and
not what triggered it. Also, it's quite hard to hit the same time using API, we are only hitting it because we are enabling
automatic backups and immediately try to do a manual backup.

Closes #302

Finally.

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>