fix: detect the old vs. new URL format correctly on workload proxying
Requests targeting workload proxies were incorrectly detected to be in the old URL format with the `p-` prefix when the instance name had a dash (`-`) in it.

Fix it by checking explicitly for the `p-` prefix in the logic.

Add a test to cover this case.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>
(cherry picked from commit dc7c2b3e3f89a9d76cdc5d23b4110948a15709dd)