fix: do not rely on `MachineStatus` updates when checking maintenance
This check in the `ClusterMachineConfigStatus` controller is incorrect:
there's no relation of the `MachineStatus` update and the machine
connectivity.
We should have another way to re-trigger reconciliation and one of the
options is the requeue after.

Make it re-queue the reconciliation limited amount of times.
It's not 100% fail-proof, but no worse than before.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>