feat: compute machineclass machine requirement (pressure)
Introduce the resource `MachinesetRequiredMachines` containing the information about how many more machines are requested from a machine set using a machine class.

Introduce the resource `MachineClassStatus`, containing the information of the total amount of required additional machines (i.e., machine pressure) on a machine set.

The pressure is only created if a machine set is using a machine class with static allocation type (not unlimited).

Related to siderolabs/omni#270.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>