chore: implement labels extractor controller for more efficient code
The new controller is generic and allows extracting labels from any
resource types just by adding the controller to the runtime.
Using this controller will reduce reconciliation count as we ignore
resource spec updates.

Use the `ResourceLabels` resource as the input in the `MachineSetNode`.
Add `MachineStatusLabelInstalled` label to make it possible for
`MachineSetNode` controller to operate only with labels.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>