feat: implement `MachineRequestSets` and support links cleanup flow
The `MachineRequestSet` creates a number of `MachineRequests`.
The machine request count is defined in the `MachineRequestSet` along
with schematic id, Talos version, provider data and provider id to use.

Also introduce `MachineRequestCleanup` controller that can initiate the
machine set node removal if the machine is added to a cluster and after
the cluster machine is gone it should tear down the link.

`MachineSetNodeController` now also reacts on the `MachineRequestStatus`
teardown: it will remove the related `MachineSetNode`.

Fixes: https://github.com/siderolabs/omni/issues/501

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>