fix: display machines in tearing down state
We display `MachineStatusLink` on the `Machines` page, which was removed
right after `MachineStatus` enters tearing down.
But the `MachineStatus` can have finalizers pending for a long time, so
the user gets false impression that the machine was already removed,
while it's not.

Make `MachineStatusLink` remain until the corresponding `MachineStatus`
exist.
Add the flag to the spec to allow displaying that the machine is
in tearing down phase in the UI.

Show machines in tearing down phase with 50% opacity and disable any
operations with such machine.

Fixes: https://github.com/siderolabs/omni/issues/516

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>