feat: support passing extra data through the siderolink join token
This extra data is used in the infra provider to add the annotation to the
`siderolink.Link` as early as possible.
Then the `Machine` controller is changed to skip the `Links` that have
annotation `omni.sidero.dev/infra-provider` and do not have the label
`omni.sidero.dev/machine-request`.
This change makes not consistent `Links` to be ignored by the system,
until the are fully populated.

Also changed the infra provider interface to take siderolink connection
params as string instead of the resource.

Fixes: https://github.com/siderolabs/omni/issues/603

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>