fix: use YAML representation in `omnictl get`
The benefit is that if `omnictl` is out of sync with the server, it will
still show correct server-side representation.

(If protobuf representation is out of sync with the server, it might
skip unknown fields).

This doesn't affect other commands, like `apply` or `cluster template`
family of functions, they still use protobuf representation.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>