fix: pin apexcharts version to 3.45.2
The most recent apexcharts version seems to have a bug and doesn't work
well with vue3-apexcharts wrapper. Size of charts is different, there's
an exception with negative svg sizes before it draws the chart for the
first time.

vue3-apexcharts seem to pull the latest version always, so pin it
explicitly to 3.45.2.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>