feat: support auto provisioned machine classes
The `MachineClass` now can operate in two different modes:

- manual, which selects the machines using the label selectors.
- auto provisioned, which can automatically create and delete machines
  when they are requested by the machine sets which are using the
  machine class.

`MachineRequest` and `MachineRequestSet` now have the provider data
field which is validated by the JSON schema which is uploaded to Omni by
the infrastructure providers.

On the UI side introduce the new components for creating the machine
classes which auto provision machines.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>