fix: get proper IP from peer metadata
Currently, if gateway meets existing X-Forwarded-For header, it will append peer address that it sees to the existing value using comma.
Our IP extraction function didn't account for that, and so it failed to parse IP and it used the original `peer.address` which
set deep below in the gRPC middleware.

This commit ensures that we try to split the string value using `,`.

Closes #668

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>