feat: display pending machine requests in the machine sets
The provision controller was changed: now it creates individual machine
request sets for each machine set which is using the machine class with
auto provision.
That allows to map machine requests to each machine set.

Talos version and initial extensions was dropped from the
autoprovisioned machine classes as Talos version is now read from the
Cluster to which the machine request is going to be added.
Extensions support is dropped as there's no way to reliably set them
without a Talos version.

Machine request set pressure resource was removed as the controller now
reads the required machine count directly from the machine sets.

Infra provider controllers were updated to report human readable status.

Introduced the new resource type `ClusterMachineRequestStatus` that is
used in the UI to show pending machine requests.

Closes: https://github.com/siderolabs/omni/issues/662

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>