feat: introduce resource for infra provider health checks
Add the ephemeral resource `InfraProviderHealthStatus` for the providers to report their health.

Refactor the infra provider state to define resource characteristics/configuration in a single place.

Add a new ephemeral underlying state to the infra provider state for it to be able to store ephemeral resources.

Improve/do fixes on resource operation checks in infra provider state.

Introduce the new `ProviderHealthStatusController` for the infra providers to report their health status.

This controller is used by the common infra provider library, but can be used standalone as well.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>