feat: push a partial config to machines in maintenance mode
Our goal is to ensure that the machines connected to Omni have always a correct machine config.

To do that, push a partial config with an `EventSinkConfig` and a `KmsgLogConfig` to the machines in maintenance mode.

If they have an existing partial config, merge the new documents with it.

Re-apply this partial config each time the machine disconnects and reconnects (siderolink public key changes), as it might mean the machine could have been rebooted and lost its in-memory partial config.

Note: we do not push a `SideroLinkConfig`, as if the machine is already connected to Omni, it has its SideroLink configured one way or the other. Additionally, we do not want to interfere/overwrite the join token machine uses to connect.
Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>