fix: properly build the search query on the Machines page
The query was changed to use `OR`, so it was concatenating label filters
to each element of the selectors array, but if the array is empty, it
was doing nothing.
Put label filters as the first element to the selectors if the array is
empty instead.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>
(cherry picked from commit ad341821e8b9c612b106be298f51abfea1ccbb3c)