fix: preserve SideroLink tunnel config on machine allocation
When a machine is connected to SideroLink over the GRPC tunnel in maintenance mode, its tunnel config gets overwritten when a machine config is applied to it.

Instead, check the link to determine the tunnel setting and preserve it on allocation.

Since we are not explicit about this setting and not always leave it to the Omni-wide configuration, rework the Omni-wide configuration to
- allow opt-in if it is disabled instance-wide
- do not allow opt-out if it is enabled instance-wide

This allows us to preserve the current behavior of "forcing" the machines without explicit configuration to use the system default.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>