fix: remove force unique token annotation from the link on wipe
The problem with the previous approach is that any consequent node
reboot will mess up the connection flow for the machines managed by the
bare metal infra provider.

So in addition to allowing to join the nodes with the different join token
for one time, reset the annotation on the link. Which makes the machine
go to the uninstalled mode once again, until it is added to a cluster.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>