fix: do not switch Siderolink GRPC tunnel mode after provisioning
Followup fix after https://github.com/siderolabs/omni/pull/976.

Change the way the global flag `--siderolink-use-grpc-tunnel` works: It now is used in the SideroLink provision API to decide whether to ignore what the machine requests in the prevision request and "force" the tunnel mode.

Additionally, now it is ignored in the generated `SideroLinkConfig` document which is pushed to the machines when they are allocated - instead, we simply check if the machine was already connected (provisioned) using a GRPC tunnel, and preserve that option. By doing that, we avoid switching the mode after the provisioning, avoiding the bug which is fixed in https://github.com/siderolabs/talos/pull/10517.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>