feat: allow force-destroying a node in booting state
Implement force-destroy logic to allow machines to be reset and removed from the cluster even if they are stuck in the booting phase. This is useful in cases like a node not being able to join to etcd, and it is known by the cluster operator that it will never be able to.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>