fix: add annotations on the `ClusterMachines` to force enable features
This is another and less intrusive approach to enable `diskQuotaSupport`
and `apidCheckExtKeyUsage` features on the nodes affected by the version
contract mismatch.

Migration drops the config patches introduced in 0.48.2 and adds
separate annotations for each feature on the `ClusterMachine` resource.

The `MachineSetStatus` controller shouldn't care about them and preserve
them during the machine updates.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>