feat: allow managing infra providers in the UI
No longer show the infra provider service accounts on the service
accounts page.
Instead, move the infra provider service account creation to the infra
providers page.
They are now always created in pair with `infra.Provider` resources.

If `infra.Provider` resource doesn't exist the provider is not allowed
to connect.
Also `infra.Provider` resource controls the lifecycle of the
`infra.ProviderStatus`, `infra.ProviderHealthStatus` and the service
account created for the provider: when the `infra.Provider` is deleted
everything is cleaned up.

`infra.ProviderStatus` and `infra.ProviderHealthStatus` are now combined
into `infra.ProviderCombinedStatus` which represents the current
connection state of the provider: if it runs or not and if it has any
errors.

`infra.ProviderCombinedStatus` resources are now shown on the new page
in the Omni UI.

`omnictl` got new commands for managing and viewing infra providers in
the system: `omnictl infraprovider [create|delete|renewkey|list]`.

Fixes: https://github.com/siderolabs/omni/issues/894

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>