fix: correctly encode exposed service redirect URL after auth
We had a URL encoding issue in the `redirect` query parameter in the new workload proxy auth/redirect flow. On the authentication page, we URL-decoded the redirect URL to be used after the authentication, but forgot to re-encode it before issuing the redirect afterwards, which submits it to the backend.

Backend then was unable to base64-decode the signed redirect URL, resulting in a bad request.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>