fix: remove deprecated controller finalizers from the machine classes
There was `MachineClassStatusController` that wasn't used for long. It
was introduced and dropped within one minor Omni version.
But if that particular Omni version was used, the controller would add
it's finalizer on the `MachineClass` resources.
Then as the controller is dropped, nothing removes the finalizer.

Introduce migration to drop that finalizer forcefully.

Fixes: https://github.com/siderolabs/omni/issues/1197

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>