chore: introduce COSI state helpers to reduce boiler plate code count
Add the new methods for `TeardownAndDestroy` and `TeardownAndDestroyAll`
that can be used in the controllers: they have 2 step flow which calls
teardown for all resources first, then calls destroy for the resources
which are ready to be destroyed.

As the method doesn't block it can be used in the controllers.
Also switch a lot of places that were using `safe.ToSlice` and
`safe.ToMap` to use `List.Pointers()` method instead.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>