fix: make cluster validations consistent between templates and UI
Move the cluster validations which do not require server-side information (access to resources) to a common place.

Use this new validator both from the server side validations and from the cluster templates validations.

This makes the validations consistent, resolving the inconsistency where cluster names (ID) were validated on templates but not by the server.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>