fix: make sure `powering on` stage is correctly set on infra machines
Power stage watcher was working by coincidence, only when waking up at the right time.
This was because it was not getting notified when a machine was allocated to a cluster.

Fix this by also getting notified on `Created` events on `ClusterMachine`s.
Also rework the watcher, make it testable and add a unit test for it.

Additionally: Store the agent mode info in machine status, and if it is in agent mode, skip attempting to call unsupported APIs.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>