fix: fix nil dereference in machine status controller
When reconciling a `MachineStatus`, we look at the `MachineStatus` field on the `MachineStatusSnapshot` resource. This field can be nil if the snapshot contains a power stage event - i.e., when the bare-metal infra provider is used, and a power status change is happened on `InfraMachineStatus`.

Use the nil-safe getter on the spec field to avoid crashing `MachineStatusController`.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>